% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img2coord-package.R
\docType{package}
\name{img2coord-package}
\alias{img2coord}
\alias{img2coord-package}
\title{img2coord: Coordinates of Points from an Image}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
