% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{img2coord}
\alias{img2coord}
\title{Get coordinates from image of points}
\usage{
img2coord(file, x_ticks, y_ticks, K, K_min = K, K_max = K)
}
\arguments{
\item{file}{Path to image.}

\item{K}{Number of points in the image.
Use \code{K_min} and \code{K_max} if you only know an interval.}

\item{K_min}{Minimum number of points in the image.}

\item{K_max}{Maximum number of points in the image.}
}
\value{
A list of coordinates. There is also an attribute "stat" which was
used to guess the number of points in the image.
}
\description{
Get coordinates from image of points
}
\examples{
# Create some image
file <- tempfile(fileext = ".png")
png(file, width = 600, height = 400)
plot(y <- c(0, runif(20), 1))
dev.off()

# Get coordinates
(coord <- img2coord(file, seq(5, 20, 5), seq(0, 1, 0.2),
                    K_min = 10, K_max = 30))
round(coord$x, 1)
plot(coord$y, y, pch = 20, cex = 1.5); abline(0, 1, col = "red")

}
