% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick.R
\name{img_scale}
\alias{img_scale}
\title{Resize an image}
\usage{
img_scale(img, scale)
}
\arguments{
\item{img}{Magick image.}

\item{scale}{Ratio to use for rescaling (the number of pixels).}
}
\description{
Resize an image
}
