% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_centers}
\alias{get_centers}
\title{Get centers by computing the barycenter of pixels of each cluster}
\usage{
get_centers(img_mat_in, points, K_opt)
}
\description{
Get centers by computing the barycenter of pixels of each cluster
}
